feat: enhance JWT token generation to include expiration time and update related API responses for improved authentication handling

+++ migrate/migrations/m_20250826.go
@@ -7,7 +7,7 @@ import (
 
 func M20250826(db *gorm.DB) *gormigrate.Gormigrate {
 	type WithdrawRecord struct {
-		LocalStatus    int `json:"local_status" gorm:"not null;default:0;index"`
+		LocalStatus    int8 `json:"local_status" gorm:"not null;default:0;index"`
 	}
 	
 	return gormigrate.New(db, gormigrate.DefaultOptions, []*gormigrate.Migration{
